# specs/002-integrate-book-kit/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Copilot Book Writer + BookKit Agents
  version: 0.0.0
  description: |
    API contract additions for integrating BookKit agent prompt selection.
servers:
  - url: http://localhost:8787
paths:
...
  /api/sessions/{sessionId}/messages:
    post:
      summary: Send a user message to a session
      parameters:
        - in: path
          name: sessionId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendMessageRequest'
      responses:
        '200':
          description: Message accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
        '400':
          description: Bad request
        '404':
          description: Unknown session
components:
  schemas:
...
    SendMessageRequest:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
          description: The user message text.
        agentId:
          type: string
          description: Optional BookKit agent/mode id to apply to this message (not persisted).